A bunch more restructuring. Getting there slowely
