+++ Sources/SpeziBluetooth/CoreBluetooth/Configuration/ServiceDescription.swift
@@ -19,14 +19,14 @@ public struct ServiceDescription {
     ///
     /// Those are the characteristics we try to discover. If empty, we discover all characteristics
     /// on a given service.
-    public let characteristics: [CBUUID] // TODO: allow nil vs. empty to discover nothing?
+    public let characteristics: Set<CBUUID> // TODO: allow nil vs. empty to discover nothing?
 
 
     /// Create a new service description.
     /// - Parameters:
     ///   - serviceId: The bluetooth service id.
     ///   - characteristics: The characteristics we expect to be present on the service.
-    public init(serviceId: CBUUID, characteristics: [CBUUID]) {
+    public init(serviceId: CBUUID, characteristics: Set<CBUUID>) {
         self.serviceId = serviceId
         self.characteristics = characteristics
     }
@@ -36,8 +36,8 @@ public struct ServiceDescription {
     /// - Parameters:
     ///   - serviceId: The bluetooth service id string.
     ///   - characteristics: The characteristics we expect to be present on the service.
-    public init(serviceId: String, characteristics: [String]) {
-        self.init(serviceId: CBUUID(string: serviceId), characteristics: characteristics.map { CBUUID(string: $0) })
+    public init(serviceId: String, characteristics: Set<String>) {
+        self.init(serviceId: CBUUID(string: serviceId), characteristics: Set(characteristics.map { CBUUID(string: $0) }))
     }
 }
