+++ Sources/SpeziBluetooth/CoreBluetooth/Model/CharacteristicLocator.swift
@@ -1,0 +1,13 @@
+//
+// This source file is part of the Stanford Spezi open-source project
+//
+// SPDX-FileCopyrightText: 2024 Stanford University and the project authors (see CONTRIBUTORS.md)
+//
+// SPDX-License-Identifier: MIT
+//
+
+
+struct CharacteristicLocator: Hashable {
+    let serviceId: CBUUID
+    let characteristicId: CBUUID
+}
