+++ Sources/SpeziBluetooth/Model/Visitor/DeviceVisitor.swift
@@ -18,13 +18,13 @@ protocol DeviceVisitor: BaseVisitor {
 
 
 extension BluetoothDevice {
-    func accept<Visitor: DeviceVisitor>(_ visitor: inout Visitor) { // TODO: any final result?
+    func accept<Visitor: DeviceVisitor>(_ visitor: inout Visitor) {
         let mirror = Mirror(reflecting: self)
         for (_, child) in mirror.children {
             if let visitable = child as? DeviceVisitable {
                 visitable.accept(&visitor)
             } else if child is ServiceVisitable {
-                // TODO: maybe some logger to catch illegal @Characteristic stuff?
+                preconditionFailure("@Characteristic declaration found in \(Self.self). @Characteristic cannot be used within the BluetoothDevice class!")
             }
         }
     }
